list and update
I renamed num to numList.
I defined a list numSelection to store the 6 randomly selected numbers
from numList.
I also changed the value of 999 to 1000 to incorporate the value 999
into the target range. I had no need to change the value of 101 as the
min value because the min value stays in the range.